refactor all the things PlaybookEdit
