+++ webapp/src/hooks/routing.ts
@@ -12,7 +12,7 @@ import {useSelector} from 'react-redux';
 import {navigateToPluginUrl, pluginUrl} from 'src/browser_routing';
 import {TEMPLATE_TITLE_KEY} from 'src/constants';
 import {Playbook} from 'src/types/playbook';
-import {tabInfo} from 'src/components/backstage/playbook_edit';
+import {tabInfo} from 'src/components/backstage/playbook_edit/playbook_edit';
 
 type PlaybooksRoutingOptions<T> = {
     urlOnly?: boolean,
